backend/requirements.txt
# Scout Leader Lesson Architect - Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Data processing
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.10.4
lxml==5.1.0

# AI
openai==1.10.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Generation
weasyprint==60.2
reportlab==4.0.9

# Weather API
pyowm==3.3.0

# Vector search
pgvector==0.2.2